+++ bcs-app/backend/bcs_web/permissions.py
@@ -76,7 +76,6 @@ class ProjectEnableBCS(BasePermission):
     主要功能:
     - 校验项目是否已经开启容器服务
     - 设置 request.project、request.ctx_project 、request.ctx_cluster，在 view 中使用
-    - 设置 request.audit_ctx，可配合 log_audit_on_view 和 log_audit 装饰器使用
     """
 
     message = "project does not enable bcs"
@@ -87,8 +86,6 @@ class ProjectEnableBCS(BasePermission):
         if project:
             request.project = project
             self._set_ctx_project_cluster(request, project.project_id, view.kwargs.get('cluster_id', ''))
-            # 设置操作审计 context
-            request.audit_ctx = AuditContext(user=request.user.username, project_id=project.project_id)
             return True
 
         return False
